requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.1
djangorestframework==3.15.2
idna==3.10
kombu==5.4.1
numpy==2.1.1
pandas==2.2.2
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13